perf(hyperdim): accelerate parallel bundling with SIMD
Optimized the parallel bundling path in HVec10240::bundle to use
architecture-specific SIMD intrinsics (AVX2/NEON) instead of scalar
word processing. Added correctness tests for NEON and consistency
tests across threshold boundaries. Resolved clippy lint regarding
unnecessary casts in tests.

+++ src/hyperdim_tests.rs
@@ -99,7 +99,7 @@ fn test_bundle_threshold_consistency() {
             }
             let mut word_res = 0u128;
             for j in 0..128 {
-                if bit_counts[j] >= threshold as i32 {
+                if bit_counts[j] >= threshold {
                     word_res |= 1u128 << j;
                 }
             }
